Migrate from Modules to standalone components
