--- apps/expenses-web/src/app/app-routing.module.ts
@@ -1,33 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { AuthGuard } from './core/auth/auth.guard';
-
-const routes: Routes = [
-  {
-    path: 'auth',
-    loadChildren: () =>
-      import('./core/auth/auth.module').then((m) => m.AuthModule),
-  },
-  {
-    path: 'user',
-    loadChildren: () =>
-      import('./shell/user/user.module').then((m) => m.UserModule),
-  },
-  {
-    path: 'features',
-    loadChildren: () =>
-      import('./features/features.module').then((m) => m.FeaturesModule),
-    canActivate: [AuthGuard],
-  },
-  {
-    path: '**',
-    redirectTo: 'features',
-    pathMatch: 'full',
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule],
-})
-export class AppRoutingModule {}
