+++ apps/expenses-web/src/app/app.component.ts
@@ -6,14 +6,17 @@ import {
   RouteConfigLoadEnd,
   RouteConfigLoadStart,
   Router,
+  RouterOutlet,
 } from '@angular/router';
 import { SpinnerService } from './shell/spinner/spinner.service';
+import { SpinnerComponent } from './shell/spinner/spinner.component';
+import { NotificationComponent } from './shell/notification/notification.component';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
-  standalone: false,
+  imports: [RouterOutlet, SpinnerComponent, NotificationComponent],
 })
 export class AppComponent implements OnInit, OnDestroy {
   private destory$ = new Subject<void>();
