+++ apps/expenses-web/src/app/app.config.ts
@@ -1,8 +1,75 @@
 import { ApplicationConfig } from '@angular/core';
 import { provideRouter } from '@angular/router';
 
-import { routes } from './app.routes';
+import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
+import { providePrimeNG } from 'primeng/config';
+import {
+  HTTP_INTERCEPTORS,
+  provideHttpClient,
+  withInterceptorsFromDi,
+  HttpClient,
+} from '@angular/common/http';
+import { AuthInterceptor } from './core/auth/auth.interceptor';
+import { BrowserModule } from '@angular/platform-browser';
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { isDevMode, importProvidersFrom } from '@angular/core';
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
+import { APP_ROUTES } from './app.routes';
+import Aura from '@primeng/themes/aura';
+import { UserEffect } from './shell/user/store/effects/user.effect';
+import { userFeature } from './shell/user/store/features/user.feature';
+import { AuthEffects } from './core/auth/store/effects/auth.effects';
+import { authFeature } from './core/auth/store/feature/auth.feature';
+import { localStorageSyncReducer } from './core/auth/store/local-storage-sync.reducer';
+
+function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http, './assets/i18n/', '.json');
+}
 
 export const appConfig: ApplicationConfig = {
-  providers: [provideRouter(routes)],
+  providers: [
+    provideRouter(APP_ROUTES),
+    importProvidersFrom(
+      BrowserModule,
+      BrowserAnimationsModule,
+      StoreModule.forRoot({}),
+      EffectsModule.forRoot([]),
+      StoreDevtoolsModule.instrument({
+        maxAge: 25, // Retains last 25 states
+        logOnly: !isDevMode(), // Restrict extension to log-only mode
+        autoPause: true, // Pauses recording actions and state changes when the extension window is not open
+        trace: false, //  If set to true, will include stack trace for every dispatched action, so you can see it in trace tab jumping directly to that part of code
+        traceLimit: 75, // maximum stack trace frames to be stored (in case trace option was provided as true)
+        connectInZone: true, // If set to true, the connection is established within the Angular zone
+      }),
+      TranslateModule.forRoot({
+        defaultLanguage: 'en',
+        loader: {
+          provide: TranslateLoader,
+          useFactory: HttpLoaderFactory,
+          deps: [HttpClient],
+        },
+      }),
+
+      StoreModule.forFeature(authFeature.name, authFeature.reducer, {
+        metaReducers: [localStorageSyncReducer],
+      }),
+      EffectsModule.forFeature([AuthEffects]),
+      StoreModule.forFeature(userFeature),
+      EffectsModule.forFeature([UserEffect]),
+    ),
+    provideAnimationsAsync(),
+    providePrimeNG({
+      theme: {
+        preset: Aura,
+      },
+    }),
+    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+    provideHttpClient(withInterceptorsFromDi()),
+  ],
 };
