--- apps/expenses-web/src/app/features/expenses/expenses-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { ExpensesComponent } from './components/expenses.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ExpensesComponent,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class ExpensesRoutingModule {}
