+++ apps/expenses-web/src/app/features/features.component.ts
@@ -1,9 +1,11 @@
 import { Component } from '@angular/core';
+import { SidebarComponent } from '../shell/sidebar/sidebar.component';
+import { RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'app-features',
   templateUrl: 'features.component.html',
   styleUrls: ['./features.component.scss'],
-  standalone: false,
+  imports: [SidebarComponent, RouterOutlet],
 })
 export class FeaturesComponent {}
