+++ apps/expenses-web/src/app/features/home/components/no-data/no-data.component.ts
@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-no-data',
   templateUrl: 'no-data.component.html',
-  standalone: false,
+  imports: [TranslateModule],
 })
 export class NoDataComponent {}
