--- apps/expenses-web/src/app/features/incomes/incomes-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { IncomesComponent } from './components/incomes.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: IncomesComponent,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class IncomesRoutingModule {}
