+++ apps/expenses-web/src/app/features/settings/components/settings.component.ts
@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
+import { AppContentWrapperComponent } from '../../../shared/components/app-content-wrapper/app-content-wrapper.component';
 
 @Component({
   selector: 'app-settings',
   templateUrl: 'settings.component.html',
   styleUrls: ['./settings.component.scss'],
-  standalone: false,
+  imports: [AppContentWrapperComponent],
 })
 export class SettingsComponent {}
