+++ apps/expenses-web/src/app/features/settings/settings.routes.ts
@@ -1,0 +1,9 @@
+import { Routes } from '@angular/router';
+import { SettingsComponent } from './components/settings.component';
+
+export const SETTINGS_ROUTES: Routes = [
+  {
+    path: '',
+    component: SettingsComponent,
+  },
+];
