+++ apps/expenses-web/src/app/shared/pipes/format-date.pipe.ts
@@ -4,7 +4,6 @@ import * as dayjs from 'dayjs';
 @Pipe({
   name: 'formatDate',
   pure: true,
-  standalone: false,
 })
 export class FormatDatePipe implements PipeTransform {
   transform(value: Date | string, format = 'ddd, DD. MMM YYYY, HH:mm'): string {
