+++ apps/expenses-web/src/app/shell/notification/notification.component.ts
@@ -4,12 +4,11 @@ import { NotificationService } from './notification.service';
 import { MessagesModule } from 'primeng/messages';
 import { MessageService } from 'primeng/api';
 import { Subject, takeUntil } from 'rxjs';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @Component({
   selector: 'app-notification',
   templateUrl: 'notification.component.html',
-  imports: [ToastModule, MessagesModule, BrowserAnimationsModule],
+  imports: [ToastModule, MessagesModule],
   providers: [MessageService],
 })
 export class NotificationComponent implements OnInit, OnDestroy {
